test: move concurrent-close TracePoint test into test/unit
test_concurrent_close_releases_native_handle_once_traced never opens a
real connection (Valkey.allocate + direct ivar sets), so it belongs in
test/unit rather than mixed into the standalone-only integration module.
Dropped the now-unneeded cluster_mode? skip along with it.
